adjusted DUSL and SRP to include abstract download links
